+++ tutorpanorama/__about__.py
@@ -1,4 +1,4 @@
 """
 Package version.
 """
-__version__ = "18.1.4"
+__version__ = "18.2.0"