+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change log
 
-## Unreleased
+## Version 18.2.0 (2025-03-26)
 - Fix: Remove PANORAMA_ADD_HEADER_LINK and replaced by PANORAMA_ADD_DASHBOARD_LINK.
 Now the Panorama link is in the learner-dashboard MFE instead of a header.